UI improvements, stream correction, updating UI on task reception and ABOUT to update UI when task is completed.. need to take a minute and step back and see what's broken.
TODO:
Replacing tasks in the tableview is broken. There should be a lookup table of where temporary tasks are (there index path, tag, whatever to get it from the tableview) so it can easily be replaced. It would be silly to have to use TLIndexPaths for this, but it may be just the thing... hope not!